Support 'artifact:version@site' notation for plugins not hosted by default update site
